+++ storybook/stories/index.js
@@ -14,6 +14,7 @@ import Explore from './Explore'
 import Profile from './Profile'
 import Settings from './Settings'
 import TopCharts from './TopCharts'
+import Podcasts from './Podcasts'
 
 const store = createStore(rootReducer)
 
@@ -31,3 +32,4 @@ storiesOf('UI Components', module)
 	.add('Profile', () => <Profile />)
 	.add('Settings', () => <Settings />)
 	.add('TopCharts', () => <TopCharts />)
+	.add('Podcasts', () => <Podcasts />)
